pyproject.toml
[tool.poetry]
name = "borderregion-kde"
version = "0.1.0"
description = ""
authors = ["micsoder <michaela.x.soderholm@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "2.0.3"
geopandas = "0.13.2"
numpy = "1.25.2"
geopy = "2.3.0"
h3 = "3.7.6"
seaborn = "0.12.2"
contextily = "1.3.0"
matplotlib = "3.7.2"
shapely = "2.0.1"
python-dotenv = "0.17.1"
invoke = "2.2.0"
pylint = "2.17.5"
scikit-learn = "1.3.0"
fiona = "1.9.4.post1"
pyarrow = "13.0.0"
pytest = "7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
